210124
데미지 처리 변경
적이나 플레이어의 hp가 0이 되면 전투가 종료되도록 수정

+++ Assets/Script/GameManager.cs
@@ -254,7 +254,7 @@ public class GameManager : MonoBehaviour
                     string monstername = streamReader.ReadLine();
                     string imagename = streamReader.ReadLine();
                     string iconname = streamReader.ReadLine();
-                    float exp = float.Parse(streamReader.ReadLine());
+                    int exp = int.Parse(streamReader.ReadLine());
                     float spd = float.Parse(streamReader.ReadLine());
                     int lv = int.Parse(streamReader.ReadLine());
                     int atk = int.Parse(streamReader.ReadLine());
